% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utilities.R
\name{dataframe_generation}
\alias{dataframe_generation}
\title{Generate a Data Frame from Inline R Formulas}
\usage{
dataframe_generation(formulas, n = 300)
}
\arguments{
\item{formulas}{Character vector of code lines, or a single multi-line string.
Lines starting with \code{"#"} are ignored.}

\item{n}{Integer. Sample size available inside the temporary environment (bound to \code{n}).}
}
\value{
A data frame with one column per symbol created by the code (excluding \code{n}).
}
\description{
Evaluates a set of R statements (one per line) inside a temporary environment
that contains \code{n} and the current session scope, then returns all created
objects (except \code{n}) as a data frame.
}
\examples{
\dontrun{
code <- "
X <- rnorm(n)
Y <- X^3 + rnorm(n, 0, 0.1)
S <- rnorm(n)
W <- rnorm(n)
Z <- X^2 + log(abs(Y)) + W^2 + rnorm(n, 0, 0.1)
"
df <- dataframe_generation(code, n = 500)
str(df)

code1 <- "
X <- rnorm(n)
Y <- X^2 + rnorm(n, 0, 0.1)
W <- rnorm(n)
Z <- log(abs(W) + 1) + rnorm(n, 0, 0.1)
"
df1 <- dataframe_generation(code1, n = 500)

code2 <- "
X <- runif(n, -pi, pi)
Y <- sin(X) + rnorm(n, 0, 0.05)
Z <- cos(Y) + rnorm(n, 0, 0.05)
"
df2 <- dataframe_generation(code2, n = 500)

code3 <- "
L <- rnorm(n)
W <- rnorm(n)
Z <- rnorm(n)
X <- L^4 + rnorm(n, 0 , 0.1)
Y <- W^5 + rnorm(n, 0, 0.1)
"
df3 <- dataframe_generation(code3, n = 500)

code4 <- "
X <- rnorm(n)
Y <- X^2 + rnorm(n, 0, 0.1)
Z <- rnorm(n, 0, 0.1)
S <- exp(abs(Z)) + rnorm(n, 0, 0.1)
"
df4 <- dataframe_generation(code4, n = 500)

code5 <- "
X <- rnorm(n)
Y <- 2*X + rnorm(n, 0, 0.1)
Z <- 1/(X+1)^2 + rnorm(n, 0, 0.1)
S <- atan(Z) + rnorm(n, 0, 0.1)
"
df5 <- dataframe_generation(code5, n = 500)
}

}
