% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{relu}
\alias{relu}
\title{ReLU (Rectified Linear Unit) activation}
\usage{
relu(x)
}
\arguments{
\item{x}{Numeric vector or matrix input.}
}
\value{
A matrix with negative values replaced by zero.
}
\description{
Sets all negative values to zero.
}
