% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid activation}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{Numeric vector or matrix input.}
}
\value{
Transformed values in (0, 1).
}
\description{
Computes the sigmoid transformation.
}
